requirements.txt
fastapi
uvicorn
requests
PyMuPDF
openai
python-dotenv